Maze Game Patch 0.7
Added a new mechanics to the game: Yellow Doors.
These doors will not open until a key is picked up (The Green Key), allowing the player to continue on their path. 

A side block outside of the game border was added to test future iterations that will remind the player they picked up a key.