Maze Game Patch 0.7.1
Modified the texture and colors to offer enhanced clarity. 
This builds on a concept that colored walls that differ from the darken blue color are Doors.
Blue doors are Doors that can be removed by picking up a corresponding color key. (Blue Keys open Blue Doors)
The Yellow door in this level can not be removed by any means. Open to being changed.

The outside box was removed to simplify the game's parameters.